requirements.txt
cffi
PyNaCl
discord.py[voice]==0.11.0
gmusicapi
youtube-dl
psycopg2
selenium
markovify
